Organization and Hierarchy
Due to not using my laptop, I've fallen out of practice with using Git for this project. This major change includes a structure for organizing suggested room allocations before enacting them, allowing for better grouping/merging of regions. It also moves lower level operations out of control of the builder, as the system should be abstracted away from the layout logic as much as possible. The system also includes a myriad of bug fixes in regards to merging/size metrics/room typing, with one exception that I believe to have tracked down, and will isolate/remove in the next bug.